A sample to show the use of GPS capabilities when available.
This sample uses the FusedLocation APIs of GMS on those
devices that have a hardware GPS built in. In those cases,
this sample provides a simple screen that shows the current
speed of the device on the watch. User can set a speed limit
and if the speed approaches that limit, it changes the color
to yellow and if it exceeds the limit, it turns red. User
can also enable recording of coordinates and when it pairs
back with the phone, this data will be synced with the phone
component of the app and user can see a track made of those
coordinates on a map on the phone.

Change-Id: I05e23b3b6274b9a94708f5d8765067400c676924