Make samples stop updating in ambient mode.
Since watch faces are now visible in ambient mode, we need to change
the way we update our watch faces. For continuously updating watch
faces like Sweep and Tilt, onDraw() should only call invalidate()
when the watch face is visible and we're in interactive mode. For
ticking watch faces like Analog and Digital, we add logic so that
the timer only runs when the watch face is visible and we're in
interactive mode.

Bug: 18592914

Change-Id: I234b42e450457ebddf583679b9bca50db29e7dd8